Split ShowToast.cs into ShowToastBehaviour.cs
